--- src/components/Player.jsx
@@ -1,7 +1,0 @@
-export function Player({name, img, ranking, alt}){
-    return <div>
-        <h4>{name}</h4>
-        <img src={img} alt={`${name}_${alt}`} />
-        <p>{ranking}</p>
-    </div>
-}
