+++ src/components/Player/Player.jsx
@@ -1,0 +1,13 @@
+//import './Player.scss'
+
+const Player = ({ name, img, ranking, alt }) => {
+    return (
+        <div>
+            <h4>{name}</h4>
+            <img src={img} alt={`${name}_${alt}`} />
+            <p>{ranking}</p>
+        </div>
+    )
+}
+
+export default Player
